+++ fdb-record-layer-lucene/src/main/java/com/apple/foundationdb/record/lucene/directory/FDBDirectoryManager.java
@@ -31,7 +31,6 @@ import com.apple.foundationdb.record.RecordCursor;
 import com.apple.foundationdb.record.ScanProperties;
 import com.apple.foundationdb.record.cursors.ChainedCursor;
 import com.apple.foundationdb.record.logging.KeyValueLogMessage;
-import com.apple.foundationdb.record.logging.LogMessageKeys;
 import com.apple.foundationdb.record.lucene.LuceneAnalyzerWrapper;
 import com.apple.foundationdb.record.lucene.LuceneIndexTypes;
 import com.apple.foundationdb.record.lucene.LuceneLogMessageKeys;
@@ -50,8 +49,6 @@ import com.apple.foundationdb.tuple.TupleHelpers;
 import org.apache.lucene.index.DirectoryReader;
 import org.apache.lucene.index.IndexReader;
 import org.apache.lucene.index.IndexWriter;
-import org.apache.lucene.index.MergeScheduler;
-import org.apache.lucene.index.MergeTrigger;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -318,14 +315,5 @@ public class FDBDirectoryManager implements AutoCloseable {
                 .stream()
                 .filter(i -> LuceneIndexTypes.LUCENE.equals(i.getType()))
                 .count());
-    }
-
-    public static String getMergeLogMessage(@Nonnull MergeScheduler.MergeSource mergeSource, @Nonnull MergeTrigger trigger,
-                                            @Nonnull IndexMaintainerState state, @Nonnull String logMessage) {
-        return KeyValueLogMessage.of(logMessage,
-                LuceneLogMessageKeys.MERGE_SOURCE, mergeSource,
-                LuceneLogMessageKeys.MERGE_TRIGGER, trigger,
-                LogMessageKeys.INDEX_NAME, state.index.getName(),
-                LogMessageKeys.INDEX_SUBSPACE, state.indexSubspace);
     }
 }