+++ fdb-record-layer-lucene/src/test/java/com/apple/foundationdb/record/lucene/LuceneIndexMaintenanceTest.java
@@ -139,6 +139,7 @@ public class LuceneIndexMaintenanceTest extends FDBRecordStoreTestBase {
             final int docCount = random.nextInt(10) + 1;
             try (FDBRecordContext context = openContext(contextProps)) {
                 schemaSetup.accept(context);
+                recordStore.getIndexDeferredMaintenanceControl().setAutoMergeDuringCommit(false);
                 for (int j = 0; j < docCount; j++) {
                     final int group = isGrouped ? random.nextInt(random.nextInt(10) + 1) : 0; // irrelevant if !isGrouped
                     final Tuple groupTuple = isGrouped ? Tuple.from(group) : Tuple.from();
@@ -200,10 +201,11 @@ public class LuceneIndexMaintenanceTest extends FDBRecordStoreTestBase {
     private Tuple saveRecords(final boolean isSynthetic, final int group, final int countInGroup, final long timestamp, final Random random) {
         var parent = TestRecordsGroupedParentChildProto.MyParentRecord.newBuilder()
                 .setGroup(group)
-                .setRecNo(1000L + countInGroup)
+                .setRecNo(1001L + countInGroup)
                 .setTimestamp(timestamp)
                 .setTextValue("A word about what I want to say")
                 .setIntValue(random.nextInt())
+                .setChildRecNo(1000L - countInGroup)
                 .build();
         Tuple primaryKey;
         if (isSynthetic) {